speech-to-text/requirements.txt
grpcio
grpcio-tools
openai-whisper